Label validation fails for interpolated values.
Error message on `terraform plan`
```
module.analytics-production-events-efs.kubernetes_persistent_volume.events: metadata.0.labels ("${local.persistent_volume_name}") a valid label must be an empty string or consist of alphanumeric characters, '-', '_' or '.', and must start and end with an alphanumeric character (e.g. 'MyValue',  or 'my_value',  or '12345', regex used for validation is '(([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])?')
```

This may be a core bug.
Removing this validation check to allow interpolation in label values.